requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
wandb>=0.16.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
tiktoken>=0.5.0
accelerate>=0.25.0
